# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A distributed async job queue system built with Go and Redis.",
        "title": "Async Job Queue API",
        "contact": {
            "name": "Kireeti Varma",
            "url": "https://github.com/kireetivar/async-job-queue"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
        "/dashboard": {
            "get": {
                "description": "Returns health statistics for all queues including pending, dead counts and pause status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitoring"
                ],
                "summary": "Queue health dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QueueStats"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs": {
            "post": {
                "description": "Create and enqueue a new job to the specified queue with a payload and priority.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Enqueue a new job",
                "parameters": [
                    {
                        "description": "Job creation payload",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateJob"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "description": "Retrieve the full details and current status of a job by its UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Job"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancel a pending or delayed job by its ID. Removes it from both active and delayed queues.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Cancel a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{id}/retry": {
            "post": {
                "description": "Reset a dead or failed job and re-enqueue it for processing. Only works on jobs with status dead or failed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Retry a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "406": {
                        "description": "Job is not in a retryable state",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/queues": {
            "get": {
                "description": "Returns a list of all known queues with their current depth and pause status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queues"
                ],
                "summary": "List all queues",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QueueInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/queues/{name}/pause": {
            "post": {
                "description": "Pause a queue so workers stop dequeuing new jobs from it. In-flight jobs will complete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queues"
                ],
                "summary": "Pause a queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/queues/{name}/resume": {
            "post": {
                "description": "Resume processing for a previously paused queue. Workers will begin dequeuing jobs again.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queues"
                ],
                "summary": "Resume a queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateJob": {
            "type": "object",
            "required": [
                "queue",
                "type"
            ],
            "properties": {
                "max_retries": {
                    "type": "integer",
                    "example": 3
                },
                "payload": {
                    "type": "object"
                },
                "priority": {
                    "type": "integer",
                    "example": 5
                },
                "queue": {
                    "type": "string",
                    "example": "emails"
                },
                "type": {
                    "type": "string",
                    "example": "send_welcome_email"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
        "api.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.Job": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_retries": {
                    "type": "integer"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "queue": {
                    "type": "string"
                },
                "retry_at": {
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "run_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "enqueued",
                        "running",
                        "completed",
                        "failed",
                        "dead",
                        "cancelled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.JobStatus"
                        }
                    ],
                    "example": "enqueued"
                },
                "type": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "models.JobStatus": {
            "description": "Status of a job (enqueued, running, completed, failed, dead, cancelled)",
            "type": "string",
            "enum": [
                "enqueued",
                "running",
                "completed",
                "failed",
                "dead",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusEnqueued",
                "StatusRunning",
                "StatusCompleted",
                "StatusFailed",
                "StatusDead",
                "StatusCancelled"
            ]
        },
        "models.QueueInfo": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "paused": {
                    "type": "boolean"
                }
            }
        },
        "models.QueueStats": {
            "type": "object",
            "properties": {
                "dead": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "paused": {
                    "type": "boolean"
                },
                "pending": {
                    "type": "integer"
                }
            }
        }
    },
    "tags": [
        {
            "description": "All operations related to creating, retrieving, cancelling, and retrying jobs",
            "name": "jobs"
        },
        {
            "description": "Operations for managing queues (list, pause, resume)",
            "name": "queues"
        },
        {
            "description": "Health checks and dashboard statistics",
            "name": "monitoring"
        }
    ]
}